Renamed files and removed string concatenation
