+++ 3574/2/commands/checkin_command.rb
