+++ 3574/2/helpers/photo_helper.rb
@@ -8,12 +8,12 @@ module PhotoHelper
   private
 
   def path_to_photo(file_id)
-    url = "#{BOT_PATH}getFile?file_id=#{file_id}"
+    url = "#{BOT_PATH}#{file_id}"
     raw_json = URI.open(url).read
     JSON.parse(raw_json, symbolize_names: true).dig(:result, :file_path)
   end
 
   def photo_for_download
-    URI.open(BOT_PATH_FOR_DOWNLOAD + session[:path_to_photo])
+    URI.open("#{BOT_PATH_FOR_DOWNLOAD}/#{session[:path_to_photo]}")
   end
 end
