+++ 3574/2/helpers/save_helper.rb
@@ -2,27 +2,27 @@ require 'fileutils'
 
 module SaveHelper
   def save_data(action)
-    info = path_for(action)
-    make_dir(info)
-    save_photo(info)
-    save_location(info)
+    path = path_for(action)
+    make_dir(path)
+    save_photo(path)
+    save_location(path)
   end
 
   private
 
-  def make_dir(info)
+  def make_dir(path)
     session[:timestamp] = Time.now.strftime('%Y-%m-%d_%H-%M-%S')
-    FileUtils.mkdir_p(info)
+    FileUtils.mkdir_p(path)
   end
 
-  def save_photo(info)
-    File.open(info + 'photo.jpg', 'wb') do |file|
+  def save_photo(path)
+    File.open("#{path} photo.jpg", 'wb') do |file|
       file << photo_for_download.read
     end
   end
 
-  def save_location(info)
-    File.open(info + 'location.txt', 'wb') do |file|
+  def save_location(path)
+    File.open("#{path} location.txt", 'wb') do |file|
       file << payload['location']
     end
   end
